put animal crossing in a folder
