+++ events/acnhEvent/AnimalGame.py
@@ -9,7 +9,7 @@ import discord
 import PIL
 import pymongo
 from discord.ext import commands, tasks
-
+from . import consts
 
 mclient = pymongo.MongoClient(config.mongoHost, username=config.mongoUser, password=config.mongoPass)
 
@@ -27,404 +27,12 @@ class AnimalGame(commands.Cog):
             769666021706432532,
         ]
 
-        self.animals = {
-            "Apollo": {
-                "image": "https://cdn.mattbsg.xyz/rns/Apollo-01.png",
-                "dialog1": "I feel like I just can’t stop doing the same things all the time. But I don’t care, PAH! So get me **{0}**.",
-                "dialog2": "Oh my-don’t bother me! What? I’m supposed to tell you to get me something? Fine, whatever, get me **{0}**, pah!",
-            },
-            "Beau": {
-                "image": "https://cdn.mattbsg.xyz/rns/Beau-01.png",
-                "dialog1": "Hey... saltlick. Think you can get me **{0}**? I can’t get up, I have a brand new bag of spicy chips with my name on them.",
-                "dialog2": "Listen, small favor, I could use **{0}**. Surely it would not be too much of a hassle, after all saltlick, you’re already moving!",
-            },
-            "Bill": {
-                "image": "https://cdn.mattbsg.xyz/rns/Bill-01.png",
-                "dialog1": "Yo! Still itching to do some fine handy work? That’s good! I hear you can find me **{0}**, so it would be great to work off those calories and get me them, quacko.",
-                "dialog2": "Ah! I'm workin' up a sweat! Do you think you could get **{0}** for me? I need a break, quacko.",
-            },
-            "Bree": {
-                "image": "https://cdn.mattbsg.xyz/rns/Bree-01.png",
-                "dialog1": "Everyone keeps insisting I have more than enough to be happy. But I’m not happy at all! And if I’ve ever learned anything, it’s that money buys happiness! So get me **{0}**, cheeseball.",
-                "dialog2": "Hey, cheeseball, heads up. I need **{0}**. I would get it myself but when you’re living my life you tend to be tired by the end of the minute.",
-            },
-            "Bunnie": {
-                "image": "https://cdn.mattbsg.xyz/rns/Bunnie-01.png",
-                "dialog1": "Hey, do you have **{0}**?  I've totally dreamed about getting one! Please let me know when you have it, tee-hee!",
-                "dialog2": "You know what would go perfectly with my dress? A **{0}**, tee-hee! Can you please fetch me one? ",
-            },
-            "Chief": {
-                "image": "https://cdn.mattbsg.xyz/rns/Chief-01.png",
-                "dialog1": "Listen, I REALLY need **{0}**, harrrumph. I would probably get it myself, but I'm too busy right now. It would be a huge help if you could find it for me. So how is it looking?",
-                "dialog2": "I can almost not believe I am asking you this, harrrumph. But I unfortunately need **{0}**, don’t let this get to your head!",
-            },
-            "Dobie": {
-                "image": "https://cdn.mattbsg.xyz/rns/Dobie-01.png",
-                "dialog1": "Hey, I've got a special favor to ask you... Here goes! I want **{0}**, ohmmm...",
-                "dialog2": "Oh, they don’t make them like they used to. What are them, you ask? See for yourself and bring me **{0}**, ohmmm...",
-            },
-            "Freya": {
-                "image": "https://cdn.mattbsg.xyz/rns/Freya-01.png",
-                "dialog1": "I hate to ask you for a favor, but could you please get me **{0}**, uff da? ",
-                "dialog2": "Will you lend me your hands and go bring over **{0}**, uff da?",
-            },
-            "Kyle": {
-                "image": "https://cdn.mattbsg.xyz/rns/Kyle-01.png",
-                "dialog1": "I’ve been busy making sure I look great before I go bug hunting right now so... could ya do me a solid, and get me **{0}**, alpha?",
-                "dialog2": "Oh hey, you again... alpha! So i really need a **{0}** for a party coming up, could you go do your thing and get me it?",
-            },
-            "Lobo": {
-                "image": "https://cdn.mattbsg.xyz/rns/Lobo-01.png",
-                "dialog1": "Ah-roooo! It's not a good time! I just lost my **{0}**, it was my favorite! I want another one! Do you think you could get me one?",
-                "dialog2": "Oh it's you. Well? Make yourself useful would ya? If anyone on this island could get me a **{0}** I am sure it’s you... so get on it... Ok? Ah-rooo!",
-            },
-            "Vesta": {
-                "image": "https://cdn.mattbsg.xyz/rns/NH-Vesta-Render.png",
-                "dialog1": "Baaaffo! Could you help me out, and get a **{0}** for me real quick? I know I can always count on you!",
-                "dialog2": "Buddy! I heard you're looking for some jobs to do around the island, and I really need to get some stuff done... help a girl out and procure me **{0}** please? Baaaffo!",
-            },
-            "Octavian": {
-                "image": "https://cdn.mattbsg.xyz/rns/Octavian-01.png",
-                "dialog1": "Heya, Sucker, you got a **{0}** I can have? It'd really help me out if you did!",
-                "dialog2": "Hey sucker! You’re sure running around a lot today, rather than making me dizzy, could ya make yourself useful and magic up a **{0}**?",
-            },
-            "Skye": {
-                "image": "https://cdn.mattbsg.xyz/rns/Skye-01.png",
-                "dialog1": "Ah... I have been running low on stuff for a while. I hope you can get me **{0}**, airmail.",
-                "dialog2": "Just a moment ago, I saw a cloud that really reminded me of **{0}**, airplane. If you find one, could you please bring it over?",
-            },
-            "Tank": {
-                "image": "https://cdn.mattbsg.xyz/rns/Tank-01.png",
-                "dialog1": "Yoo, I've been training so hard I forgot I needed a **{0}** for my cool down... would you mind getting me some while I finish my last lap? kerPOW!",
-                "dialog2": "Wassup! I just finished my jog, and I was wondering, any ideas where i could get a **{0}**, if you can, hit me up, you know where I'll be, kerPOW!",
-            },
-            "Vivian": {
-                "image": "https://cdn.mattbsg.xyz/rns/Vivian-01.png",
-                "dialog1": "Oh hey! Could you do me, the SOON TO BE ULTRA FAMOUS Vivian a favor? Go and fetch me a **{0}**, piffle!",
-                "dialog2": "Sooo, you prefer stuffing your face to taking a nap, eh? You know what, would you mind making yourself useful and bring me **{0}**, piffle?",
-            },
-            "Whitney": {
-                "image": "https://cdn.mattbsg.xyz/rns/Whitney-01.png",
-                "dialog1": "Oh, you're truly priceless, you know that? Would you mind bringing me a **{0}**, snappy?",
-                "dialog2": "I hate to admit it but you have some skill. So your my only beacon of hope to get **{0}**, snappy.",
-            },
-            "Zucker": {
-                "image": "https://cdn.mattbsg.xyz/rns/Zucker-01.png",
-                "dialog1": "Hey! Bloop! The bugs tell me I really need to get a **{0}** right now, could you help a fella out?",
-                "dialog2": "Haven't seen you in at least a second! Umm I got a favor to ask, bloop, I heard you're the person to ask to get hold of **{0}**, could you hook me up?",
-            },
-        }
-        self.fruits = {
-            "apple": "https://cdn.mattbsg.xyz/rns/apple.png",
-            "orange": "https://cdn.mattbsg.xyz/rns/orange.png",
-            "peach": "https://cdn.mattbsg.xyz/rns/peach.png",
-            "pear": "https://cdn.mattbsg.xyz/rns/pear.png",
-            "cherry": "https://cdn.mattbsg.xyz/rns/cherries.png",
-            "turnip": "https://cdn.mattbsg.xyz/rns/turnip.png",
-        }
-        self.fish = {
-            "black-bass": {
-                "name": "Black Bass",
-                "image": "https://cdn.mattbsg.xyz/rns/black-bass.png",
-                "value": 100,
-                "weight": 4,
-                "pun": "Turn up the beat!",
-            },
-            "carp": {
-                "name": "Carp",
-                "image": "https://cdn.mattbsg.xyz/rns/carp.png",
-                "value": 50,
-                "weight": 5,
-                "pun": "Sir, this is the HOV lane.",
-            },
-            "crucian-carp": {
-                "name": "Crucian Carp",
-                "image": "https://cdn.mattbsg.xyz/rns/crucian-carp.png",
-                "value": 50,
-                "weight": 5,
-                "pun": "At least it's not a crustacean.",
-            },
-            "dab": {
-                "name": "Dab",
-                "image": "https://cdn.mattbsg.xyz/rns/dab.png",
-                "value": 75,
-                "weight": 5,
-                "pun": "Dabbing since 2014.",
-            },
-            "freshwater-goby": {
-                "name": "Freshwater Goby",
-                "image": "https://cdn.mattbsg.xyz/rns/freshwater-goby.png",
-                "value": 150,
-                "weight": 4,
-                "pun": "Gotta go-by some more bait.",
-            },
-            "loach": {
-                "name": "Loach",
-                "image": "https://cdn.mattbsg.xyz/rns/loach.png",
-                "value": 100,
-                "weight": 4,
-                "pun": "Stop! Do not approach!",
-            },
-            "ocean-sunfish": {
-                "name": "Ocean Sunfish",
-                "image": "https://cdn.mattbsg.xyz/rns/ocean-sunfish.png",
-                "value": 200,
-                "weight": 3,
-                "pun": "Maybe tonight I'll find an ocean moon-fish!",
-            },
-            "olive-flounder": {
-                "name": "Olive Flounder",
-                "image": "https://cdn.mattbsg.xyz/rns/olive-flounder.png",
-                "value": 150,
-                "weight": 4,
-                "pun": "I found 'er",
-            },
-            "red-snapper": {
-                "name": "Red Snapper",
-                "image": "https://cdn.mattbsg.xyz/rns/red-snapper.png",
-                "value": 700,
-                "weight": 2,
-                "pun": "You've been caught red handed!",
-            },
-            "sea-bass": {
-                "name": "Sea Bass",
-                "image": "https://cdn.mattbsg.xyz/rns/sea-bass.png",
-                "value": 150,
-                "weight": 4,
-                "pun": "Best music unda da sea.",
-            },
-            "sea-butterfly": {
-                "name": "Sea Butterfly",
-                "image": "https://cdn.mattbsg.xyz/rns/sea-butterfly.png",
-                "value": 300,
-                "weight": 3,
-                "pun": "Wait, can this thing fly?",
-            },
-            "shark": {
-                "name": "Shark",
-                "image": "https://cdn.mattbsg.xyz/rns/shark.png",
-                "value": 1500,
-                "weight": 1,
-                "pun": "I finn-a-ly caught one!",
-            },
-            "squid": {
-                "name": "Squid",
-                "image": "https://cdn.mattbsg.xyz/rns/squid.png",
-                "value": 175,
-                "weight": 4,
-                "pun": "Ink-sightful!",
-            },
-            "tadpole": {
-                "name": "Tadpole",
-                "image": "https://cdn.mattbsg.xyz/rns/tadpole.png",
-                "value": 25,
-                "weight": 5,
-                "pun": "I have a few questions. Can you answer my tad poll?",
-            },
-            "water-egg": {
-                "name": "Water Egg",
-                "image": "https://cdn.mattbsg.xyz/rns/water-egg.png",
-                "value": 50,
-                "weight": 3,
-                "pun": "Water you doing in there little buddy?",
-            },
-            "yellow-perch": {
-                "name": "Yellow Perch",
-                "image": "https://cdn.mattbsg.xyz/rns/yellow-perch.png",
-                "value": 50,
-                "weight": 5,
-                "pun": "Take a seat if you are feeling mellow.",
-            },
-        }
-        self.bugs = {
-            "butterfly": {
-                "name": "Butterfly",
-                "image": "https://cdn.mattbsg.xyz/rns/common-butterfly.png",
-                "value": 100,
-                "weight": 4,
-                "pun": "It better fly.",
-            },
-            "hermit-crab": {
-                "name": "Hermit Crab",
-                "image": "https://cdn.mattbsg.xyz/rns/hermit-crab.png",
-                "value": 300,
-                "weight": 3,
-                "pun": "The laziest type of crab.",
-            },
-            "leaf-egg": {
-                "name": "Leaf Egg",
-                "image": "https://cdn.mattbsg.xyz/rns/leaf-egg.png",
-                "value": 50,
-                "weight": 2,
-                "pun": "Wait, thats not a bug?!",
-            },
-            "moth": {
-                "name": "Moth",
-                "image": "https://cdn.mattbsg.xyz/rns/moth.png",
-                "value": 100,
-                "weight": 5,
-                "pun": "Let there be light!",
-            },
-            "pill-bug": {
-                "name": "Pill Bug",
-                "image": "https://cdn.mattbsg.xyz/rns/pill-bug.png",
-                "value": 100,
-                "weight": 5,
-                "pun": "Ever hear of a rollie polly?",
-            },
-            "spider": {
-                "name": "Spider",
-                "image": "https://cdn.mattbsg.xyz/rns/spider.png",
-                "value": 200,
-                "weight": 4,
-                "pun": "Eight legs too many.",
-            },
-            "tarantula": {
-                "name": "Tarantula",
-                "image": "https://cdn.mattbsg.xyz/rns/tarantula.png",
-                "value": 900,
-                "weight": 1,
-                "pun": "Almost as hairy as my uncle (who works at Nintendo).",
-            },
-            "wharf-roach": {
-                "name": "Wharf Roach",
-                "image": "https://cdn.mattbsg.xyz/rns/wharf-roach.png",
-                "value": 100,
-                "weight": 5,
-                "pun": "Not as scary as your coach.",
-            },
-        }
-        self.items = {
-            "bait": {
-                "name": "Bait",
-                "image": "https://cdn.mattbsg.xyz/rns/bait.png",
-                "value": 100,
-            },
-            "bells": {
-                "name": "Bells",
-                "image": "https://cdn.mattbsg.xyz/rns/bells.png",
-                "value": 0,
-            },
-            "stick": {
-                "name": "Stick",
-                "image": "https://cdn.mattbsg.xyz/rns/crafting-item-1.png",
-                "value": 25,
-            },
-            "iron-nugget": {
-                "name": "Iron Nugget",
-                "image": "https://cdn.mattbsg.xyz/rns/crafting-item-2.png",
-                "value": 500,
-            },
-            "clay": {
-                "name": "Clay",
-                "image": "https://cdn.mattbsg.xyz/rns/crafting-item-5.png",
-                "value": 50,
-            },
-            "stone": {
-                "name": "Stone",
-                "image": "https://cdn.mattbsg.xyz/rns/crafting-item-6.png",
-                "value": 25,
-            },
-            "diy": {
-                "name": "DIY Recipe",
-                "image": "https://cdn.mattbsg.xyz/rns/diy.png",
-                "value": 0,
-            },
-            "furniture": {
-                "name": "Furniture",
-                "image": "https://cdn.mattbsg.xyz/rns/furniture.png",
-                "value": 0,
-            },
-            "hat": {
-                "name": "Hat",
-                "image": "https://cdn.mattbsg.xyz/rns/hat.png",
-                "value": 0,
-            },
-            "present": {
-                "name": "Present",
-                "image": "https://cdn.mattbsg.xyz/rns/present.png",
-                "value": 0,
-            },
-            "shell": {
-                "name": "Shell",
-                "image": "https://cdn.mattbsg.xyz/rns/shell.png",
-                "value": 20,
-            },
-            "conch": {
-                "name": "Conch",
-                "image": "https://cdn.mattbsg.xyz/rns/conch.png",
-                "value": 30,
-            },
-            "cowrie": {
-                "name": "Cowrie",
-                "image": "https://cdn.mattbsg.xyz/rns/cowrie.png",
-                "value": 25,
-            },
-            "coral": {
-                "name": "Coral",
-                "image": "https://cdn.mattbsg.xyz/rns/coral.png",
-                "value": 250,
-            },
-            "sand-dollar": {
-                "name": "Sand Dollar",
-                "image": "https://cdn.mattbsg.xyz/rns/sand-dollar.png",
-                "value": 10,
-            },
-            "cherry-blossom": {
-                "name": "Cherry Blossom",
-                "image": "https://cdn.mattbsg.xyz/rns/cherry-blossom-petal.png",
-                "value": 200,
-            },
-            "wood": {
-                "name": "Wood",
-                "image": "https://cdn.mattbsg.xyz/rns/crafting-item-3.png",
-                "value": 150,
-            },
-            "wood-egg": {
-                "name": "Wood Egg",
-                "image": "https://cdn.mattbsg.xyz/rns/wood-egg.png",
-                "value": 50,
-            },
-            "tree": {
-                "name": "Tree",
-                "image": "https://cdn.mattbsg.xyz/rns/tree-2b.png",
-                "value": 0,
-            },
-        }
-        self.recipes = {
-            "bed": {
-                "wood-egg": ("items", 6),
-                "wood": ("items", 20),
-                "clay": ("items", 8),
-                "iron-nugget": ("items", 8),
-            },
-            "arch": {
-                "leaf-egg": ("bugs", 2),
-                "wood": ("items", 6),
-                "cherry-blossom": ("items", 6),
-                "stone": ("items", 6),
-            },
-            "vanity": {
-                "water-egg": ("fish", 10),
-                "wood-egg": ("items", 4),
-                "wood": ("items", 18),
-                "clay": ("items", 4),
-            },
-            "wardrobe": {
-                "leaf-egg": ("bugs", 2),
-                "wood": ("items", 16),
-                "iron-nugget": ("items", 4),
-            },
-            "wreath": {
-                "leaf-egg": ("bugs", 4),
-                "water-egg": ("fish", 4),
-                "wood-egg": ("items", 4),
-                "stick": ("items", 22),
-                "stone": ("items", 6),
-                "clay": ("items", 2),
-                "cherry-blossom": ("items", 8),
-            },
-        }
+        self.animals = consts.animals
+        self.fruits = consts.fruits
+        self.fish = consts.fish
+        self.bugs = consts.bugs
+        self.items = consts.items
+        self.recipes = consts.recipes
         self.rarity = {
             1: "ultra rare",
             2: "rare",
@@ -2141,13 +1749,3 @@ class AnimalGame(commands.Cog):
             embed.description = f'No one caught the **{self.rarity[self.bugs[catch]["weight"]]} {self.bugs[catch]["name"]}** in time, it got away!'
 
         await gameMessage.edit(embed=embed)
-
-
-def setup(bot):
-    bot.add_cog(AnimalGame(bot))
-    logging.info("[Extension] Animal Crossing Event module loaded")
-
-
-def teardown(bot):
-    bot.remove_cog("AnimalGame")
-    logging.info("[Extension] Animal Crossing Event module unloaded")
